requirements.txt
torch
numpy
Pillow>=11.0.0
google-generativeai
google-genai